+++ react/country/KOR.js
@@ -82,6 +82,9 @@ export default {
     },
   ],
   geolocation: {
+    receiverName: {
+      required: true,
+    },
     postalCode: {
       valueIn: 'long_name',
       types: ['postal_code'],