+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Make `receiverName` required while using geolocation input.
+
 ## [3.13.0] - 2020-10-01
 
 ### Added